requirements.txt
# Core Dependencies
pydantic>=2.0.0
openai>=1.0.0
rich>=10.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tiktoken>=0.3.0

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1

# Note: sqlite3 and asyncio are part of Python standard library, no need to include them